Improve resolver constructor to provide initial handlers :lollipop:
