+++ index.js
@@ -2,9 +2,8 @@ const Resolver = require('./src/resolver');
 const handlers = require('./src/handlers');
 
 const getDefaultResolver = (dirname, parent) => {
-  const _resolver = new Resolver(parent);
   const folder = dirname || process.cwd();
-  _resolver.use({
+  return new Resolver(parent, {
     path: handlers.path(folder),
     file: handlers.file(folder),
     base64: handlers.base64(),
@@ -12,7 +11,6 @@ const getDefaultResolver = (dirname, parent) => {
     require: handlers.require(folder),
     exec: handlers.exec(folder)
   });
-  return _resolver;
 };
 
 module.exports = {
