+++ test/resolver.resolve.callback.test.js
@@ -11,115 +11,6 @@ function bar(input, cb) {
   setImmediate(cb.bind(null, null, `${input}_bar`));
 }
 
-test('resolver', t => {
-  const resolver = protocall.create();
-  t.is(typeof resolver, 'object');
-  t.is(typeof resolver.use, 'function');
-  t.is(typeof resolver.resolve, 'function');
-  t.is(typeof resolver.resolveFile, 'function');
-});
-
-test('use', t => {
-  const resolver = protocall.create();
-  resolver.use('foo', foo);
-
-  t.is(typeof resolver._handlers['foo'], 'object');
-  t.is(resolver._handlers['foo'].length, 1);
-});
-
-test('unuse', t => {
-  function handlerA() {
-    // noop
-  }
-
-  function handlerB() {
-    // noop
-  }
-
-  function handlerC() {
-    // noop
-  }
-
-  const resolver = protocall.create();
-  const unuseA = resolver.use('handlerA', handlerA);
-  const unuseB = resolver.use('handlerB', handlerB);
-  const unuseC = resolver.use('handlerC', handlerC);
-
-  t.is(typeof unuseA, 'function');
-  t.is(typeof unuseB, 'function');
-  t.is(typeof unuseC, 'function');
-
-  t.is(typeof resolver._handlers['handlerA'], 'object');
-  t.is(typeof resolver._handlers['handlerB'], 'object');
-  t.is(typeof resolver._handlers['handlerC'], 'object');
-
-  t.is(resolver._handlers['handlerA'].length, 1);
-  t.is(resolver._handlers['handlerB'].length, 1);
-  t.is(resolver._handlers['handlerC'].length, 1);
-
-  unuseA();
-
-  t.is(resolver._handlers['handlerA'].length, 0);
-  t.is(resolver._handlers['handlerB'].length, 1);
-  t.is(resolver._handlers['handlerC'].length, 1);
-
-  unuseC();
-
-  t.is(resolver._handlers['handlerA'].length, 0);
-  t.is(resolver._handlers['handlerB'].length, 1);
-  t.is(resolver._handlers['handlerC'].length, 0);
-
-  unuseB();
-
-  t.is(resolver._handlers['handlerA'].length, 0);
-  t.is(resolver._handlers['handlerB'].length, 0);
-  t.is(resolver._handlers['handlerC'].length, 0);
-});
-
-test('unuse stack', t => {
-  const name = 'custom';
-  function customA() {
-    // noop
-  }
-
-  function customB() {
-    // noop
-  }
-
-  function customC() {
-    // noop
-  }
-
-  const resolver = protocall.create();
-  const unuseA = resolver.use(name, customA);
-  const unuseB = resolver.use(name, customB);
-  const unuseC = resolver.use(name, customC);
-
-  t.is(typeof unuseA, 'function');
-  t.is(typeof unuseB, 'function');
-  t.is(typeof unuseC, 'function');
-
-  t.is(typeof resolver._handlers[name], 'object');
-  t.is(resolver._handlers[name].length, 3);
-
-  const removedA = unuseA();
-  t.is(removedA, customA);
-  t.is(`${name}A`, customA.name);
-  t.is(resolver._handlers[name].length, 2);
-
-  t.is(unuseA(), undefined);
-
-  const removedB = unuseB();
-  t.is(removedB, customB);
-  t.is(`${name}B`, customB.name);
-  t.is(resolver._handlers[name].length, 1);
-
-  const removedC = unuseC();
-  t.is(removedC, customC);
-  t.is(`${name}C`, customC.name);
-  t.is(resolver._handlers[name].length, 0);
-});
-
 test.cb('resolve', t => {
   const resolver = protocall.create();
   resolver.use('foo', foo);
